Move the annotation test execution logic from AnnotationsTestsBase to a separate util class AnnotationTestExecutor and move FileManager to TestFileDiscovery
